improved plotting

+++ mppi_ros/scripts/record_data.py
@@ -5,23 +5,31 @@ import rospy
 from rospkg import RosPack
 import uuid
 import pandas as pd
+import numpy as np
 from mppi_ros.msg import Data
+from scipy.ndimage.filters import gaussian_filter1d
+
+# To make sure timestamps are uniform through experiments
+from scipy.interpolate import interp1d
 
 
 class DataRecorder:
     def __init__(self, experiment_id = "exp"):
-        self.data_dict = {"experiment_id": experiment_id + str(uuid.uuid4()),
-                          "horizon": 0.0,
-                          "nr_rollouts": 0,
-                          "filter_window": 0,
-                          "tree_search": False,
-                          "stage_cost": [[]],
-                          "weights": [[]],
-                          "step_count": [[]]}
+        self.data_dict = {"id": [],
+                          "index": [],
+                          "horizon": [],
+                          "nr_rollouts": [],
+                          "filter_window": [],
+                          "tree_search": [],
+                          "stage_cost": [],
+                          "effective_samples": [],
+                          "time": []}
         self.first_call = True
         self.data_subscriber = rospy.Subscriber("/mppi_data", Data, self.data_callback, queue_size=10)
-        self.step_count = 0
         self.initial_step_count = 0
+        self.initial_time = 0
+        self.idx = 0
+        self.experiment_id = experiment_id  + str(uuid.uuid4())
 
         self.csv_file = os.path.join(RosPack().get_path("mppi_ros"), "log", "record.csv")
         rospy.loginfo("Writing to {}".format(self.csv_file))
@@ -30,22 +38,63 @@ class DataRecorder:
         if len(data.weights.array) == 0:
             return
 
+        current_time = rospy.get_rostime().to_sec()
         if self.first_call:
-            rospy.loginfo("Received first message: horizon is: {}".format(data.config.horizon))
-            self.data_dict["horizon"] = data.config.horizon
-            self.data_dict["nr_rollouts"] = data.config.nr_rollouts
-            self.data_dict["filter_window"] = data.config.filter_window
-            self.data_dict["tree_search"] = data.config.tree_search
             self.initial_step_count = data.step_count
+            self.initial_time = current_time
             self.first_call = False
 
-        self.data_dict["stage_cost"][0].append(data.stage_cost)
-        self.data_dict["weights"][0].append(data.weights.array)
-        self.data_dict["step_count"][0].append(data.step_count - self.initial_step_count)
+        self.data_dict["id"].append(self.experiment_id)
+        self.data_dict["index"].append(self.idx)
+        self.data_dict["horizon"].append(data.config.horizon)
+        self.data_dict["nr_rollouts"].append(data.config.nr_rollouts)
+        self.data_dict["filter_window"].append(data.config.filter_window)
+        self.data_dict["tree_search"].append(data.config.tree_search)
+        self.data_dict["stage_cost"].append(data.stage_cost)
+        self.data_dict["time"].append(current_time - self.initial_time)
+        self.idx += 1
+
+        effective_samples = 1.0 / (len(data.weights.array) * np.square(np.asarray(data.weights.array)).sum())
+        self.data_dict["effective_samples"].append(effective_samples)
+
+    @staticmethod
+    def resize_array(l, new_size):
+        assert isinstance(l, list)
+        if len(l) > new_size:
+            l = l[:new_size]
+        elif len(l) < new_size:
+            size = len(l)
+            l.extend([l[-1]] * (new_size - size))
+
+    def postprocess(self):
+        # Filtering of effective samples
+        self.data_dict['effective_samples'] = gaussian_filter1d(self.data_dict['effective_samples'], sigma=2)
+
+        # Resampling of cost and timestamps: data needs to be aligned for aggregation
+        dt = 0.01
+        t_min = min(self.data_dict['time'])
+        t_max = max(self.data_dict['time'])
+        t_old = self.data_dict['time']
+        t_new = np.round(np.arange(t_min, t_max, step=dt), decimals=3)
+        self.data_dict['time'] = t_new
+
+        cost_int_fun = interp1d(t_old, self.data_dict['stage_cost'], fill_value='extrapolate')
+        self.data_dict['stage_cost'] = cost_int_fun(t_new)
+
+        samples_int_fun = interp1d(t_old, self.data_dict['effective_samples'], fill_value='extrapolate')
+        self.data_dict['effective_samples'] = samples_int_fun(t_new)
+
+        for key, value in self.data_dict.items():
+            rospy.loginfo("{} {}".format(key, value))
+            if key not in ('time', 'stage_cost', 'effective_samples'):
+                self.resize_array(value, len(self.data_dict['time']))
 
     def save(self):
         df = pd.DataFrame(self.data_dict)
-        df.to_csv(self.csv_file, mode='a', header=False)
+        if os.path.isfile(self.csv_file):
+            df.to_csv(self.csv_file, mode='a', header=False)
+        else:
+            df.to_csv(self.csv_file)
 
 
 if __name__ == "__main__":
@@ -53,4 +102,6 @@ if __name__ == "__main__":
     experiment_id = rospy.get_param("~experiment_id")
     recorder = DataRecorder(experiment_id)
     rospy.spin()
+
+    recorder.postprocess()
     recorder.save()
